+++ api/tests/integration/domain/usecases/create-organizations-with-tags-and-target-profiles_test.js
@@ -6,12 +6,11 @@ import {
   TargetProfileInvalidError,
 } from '../../../../lib/domain/errors.js';
 import { Membership } from '../../../../lib/domain/models/Membership.js';
-import * as organizationInvitationService from '../../../../lib/domain/services/organization-invitation-service.js';
+import { organizationInvitationService } from '../../../../src/team/domain/services/organization-invitation.service.js';
 import { createOrganizationsWithTagsAndTargetProfiles } from '../../../../lib/domain/usecases/create-organizations-with-tags-and-target-profiles.js';
 import * as organizationValidator from '../../../../lib/domain/validators/organization-with-tags-and-target-profiles-script.js';
 import { DomainTransaction as domainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import * as dataProtectionOfficerRepository from '../../../../lib/infrastructure/repositories/data-protection-officer-repository.js';
-import * as organizationInvitationRepository from '../../../../lib/infrastructure/repositories/organization-invitation-repository.js';
 import * as organizationTagRepository from '../../../../lib/infrastructure/repositories/organization-tag-repository.js';
 import * as tagRepository from '../../../../lib/infrastructure/repositories/tag-repository.js';
 import * as targetProfileShareRepository from '../../../../lib/infrastructure/repositories/target-profile-share-repository.js';
