+++ api/tests/team/integration/domain/services/organization-invitation.service.test.js
@@ -1,16 +1,16 @@
 import _ from 'lodash';
 
-import { SendingEmailError } from '../../../../lib/domain/errors.js';
-import { EmailingAttempt } from '../../../../lib/domain/models/EmailingAttempt.js';
-import { Membership } from '../../../../lib/domain/models/Membership.js';
-import { OrganizationInvitation } from '../../../../lib/domain/models/OrganizationInvitation.js';
-import { mailService } from '../../../../lib/domain/services/mail-service.js';
-import { createOrUpdateOrganizationInvitation } from '../../../../lib/domain/services/organization-invitation-service.js';
-import * as organizationInvitationRepository from '../../../../lib/infrastructure/repositories/organization-invitation-repository.js';
-import * as organizationRepository from '../../../../src/shared/infrastructure/repositories/organization-repository.js';
-import { catchErr, databaseBuilder, expect, sinon } from '../../../test-helper.js';
+import { SendingEmailError } from '../../../../../lib/domain/errors.js';
+import { EmailingAttempt } from '../../../../../lib/domain/models/EmailingAttempt.js';
+import { Membership } from '../../../../../lib/domain/models/Membership.js';
+import { OrganizationInvitation } from '../../../../../lib/domain/models/OrganizationInvitation.js';
+import { mailService } from '../../../../../lib/domain/services/mail-service.js';
+import * as organizationInvitationRepository from '../../../../../lib/infrastructure/repositories/organization-invitation-repository.js';
+import * as organizationRepository from '../../../../../src/shared/infrastructure/repositories/organization-repository.js';
+import { organizationInvitationService } from '../../../../../src/team/domain/services/organization-invitation.service.js';
+import { catchErr, databaseBuilder, expect, sinon } from '../../../../test-helper.js';
 
-describe('Integration | Service | Organization-Invitation Service', function () {
+describe('Integration | Team | Domain | Service | organization-invitation', function () {
   describe('#createOrUpdateOrganizationInvitation', function () {
     let clock;
     const now = new Date('2021-01-02');
@@ -38,7 +38,7 @@ describe('Integration | Service | Organization-Invitation Service', function () 
       };
 
       // when
-      const result = await createOrUpdateOrganizationInvitation({
+      const result = await organizationInvitationService.createOrUpdateOrganizationInvitation({
         organizationId,
         email,
         role,
@@ -61,7 +61,7 @@ describe('Integration | Service | Organization-Invitation Service', function () 
       await databaseBuilder.commit();
 
       // when
-      const result = await createOrUpdateOrganizationInvitation({
+      const result = await organizationInvitationService.createOrUpdateOrganizationInvitation({
         organizationId: organizationInvitation.organizationId,
         email: organizationInvitation.email,
         organizationRepository,
@@ -90,7 +90,7 @@ describe('Integration | Service | Organization-Invitation Service', function () 
       mailService.sendOrganizationInvitationEmail.resolves(mailerResponse);
 
       // when
-      const result = await catchErr(createOrUpdateOrganizationInvitation)({
+      const result = await catchErr(organizationInvitationService.createOrUpdateOrganizationInvitation)({
         organizationId: organizationInvitation.organizationId,
         email,
         organizationRepository,
